Add transport layer abstraction
